tests/end2end: migrate tests to safe server code: UC09
Relates to #925
