--- tests/end2end/document/UC09_delete_section/UC09_T01_delete_section/UC09_T01_delete_section.py
@@ -1,55 +1,0 @@
-import filecmp
-import os
-import shutil
-
-from selenium.webdriver.common.by import By
-from seleniumbase import BaseCase
-
-from tests.end2end.server import SDocTestServer
-
-path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
-
-
-class Test_UC09_T01_DeleteSection(BaseCase):
-    def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
-
-        test_server = SDocTestServer.create(path_to_sandbox)
-        shutil.copyfile(
-            os.path.join(path_to_this_test_file_folder, "document.sdoc"),
-            os.path.join(path_to_sandbox, "document.sdoc"),
-        )
-
-        test_server.run()
-
-        self.open(test_server.get_host_and_port())
-
-        self.assert_text("Document 1")
-        self.assert_text("PROJECT INDEX")
-
-        self.click_xpath('//*[@data-testid="tree-file-link"]')
-
-        self.assert_text_visible("First section")
-
-        self.hover_and_click(
-            hover_selector="(//sdoc-node)[2]",
-            click_selector=(
-                '(//sdoc-node)[2]//*[@data-testid="node-delete-action"]'
-            ),
-            hover_by=By.XPATH,
-            click_by=By.XPATH,
-        )
-
-        self.assert_text_not_visible("First section")
-
-        # TODO: Assert that the TOC is also updated.
-
-        assert os.path.exists(os.path.join(path_to_sandbox, "document.sdoc"))
-        assert filecmp.cmp(
-            os.path.join(path_to_sandbox, "document.sdoc"),
-            os.path.join(
-                path_to_this_test_file_folder, "document.expected.sdoc"
-            ),
-        )
