+++ tests/end2end/document/UC09_delete_section/UC09_T01_delete_section/input/document.sdoc
